refactor(types): split Component & Classification into own modules
Component, CreateComponentParams, and UpdateComponentParams move to
src/types/component.rs; Classification and ClassificationProduct move to
src/types/classification.rs, each with a sibling test file. product.rs now
imports Component for the Product struct. The types/mod.rs facade re-exports
the relocated types so no consumer changes.

Aligns the type layer with the per-resource module pattern already used by
client/, commands/, cli/, and output/.
