+++ src/types/component.rs
@@ -1,0 +1,39 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Serialize, Deserialize)]
+#[non_exhaustive]
+pub struct Component {
+    pub id: u64,
+    #[serde(default)]
+    pub name: String,
+    #[serde(default)]
+    pub description: String,
+    #[serde(default)]
+    pub is_active: bool,
+    #[serde(default)]
+    pub default_assignee: Option<String>,
+}
+
+#[derive(Debug, Serialize)]
+#[non_exhaustive]
+pub struct CreateComponentParams {
+    pub product: String,
+    pub name: String,
+    pub description: String,
+    pub default_assignee: String,
+}
+
+#[derive(Debug, Default, Serialize)]
+#[non_exhaustive]
+pub struct UpdateComponentParams {
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub name: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub description: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub default_assignee: Option<String>,
+}
+
+#[cfg(test)]
+#[path = "component_tests.rs"]
+mod tests;
