+++ src/types/mod.rs
@@ -1,7 +1,9 @@
 mod attachment;
 mod bug;
+mod classification;
 mod comment;
 mod common;
+mod component;
 mod group;
 mod product;
 mod user;
@@ -12,15 +14,16 @@ pub use bug::{
     FieldValue, HistoryEntry, IdListUpdate, NegationOp, Overrides, QueryKind, SavedQuery,
     SearchParams, StatusTransition, StringListUpdate, UpdateBugParams, FIELD_MAPPINGS,
 };
+pub use classification::{Classification, ClassificationProduct};
 pub use comment::{Comment, UpdateCommentTagsParams};
 pub use common::{
     ApiMode, AuthMethod, ExtensionInfo, FlagStatus, FlagUpdate, OutputFormat, ServerExtensions,
     ServerInfoResponse, ServerVersion, SortDirection,
 };
+pub use component::{Component, CreateComponentParams, UpdateComponentParams};
 pub use group::{CreateGroupParams, GroupInfo, GroupMember, UpdateGroupParams};
 pub use product::{
-    Classification, ClassificationProduct, Component, CreateComponentParams, CreateProductParams,
-    Milestone, Product, ProductListType, UpdateComponentParams, UpdateProductParams, Version,
+    CreateProductParams, Milestone, Product, ProductListType, UpdateProductParams, Version,
 };
 pub use user::{BugzillaUser, CreateUserParams, UpdateUserParams, UserGroup, WhoamiResponse};
