+++ src/types/product.rs
@@ -1,5 +1,7 @@
 use serde::{Deserialize, Serialize};
 
+use crate::types::component::Component;
+
 #[derive(Debug, Serialize, Deserialize)]
 #[non_exhaustive]
 pub struct Product {
@@ -16,20 +18,6 @@ pub struct Product {
     pub versions: Vec<Version>,
     #[serde(default)]
     pub milestones: Vec<Milestone>,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-#[non_exhaustive]
-pub struct Component {
-    pub id: u64,
-    #[serde(default)]
-    pub name: String,
-    #[serde(default)]
-    pub description: String,
-    #[serde(default)]
-    pub is_active: bool,
-    #[serde(default)]
-    pub default_assignee: Option<String>,
 }
 
 #[derive(Debug, Serialize, Deserialize)]
@@ -74,50 +62,6 @@ pub struct UpdateProductParams {
     pub default_milestone: Option<String>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub is_open: Option<bool>,
-}
-
-#[derive(Debug, Serialize)]
-#[non_exhaustive]
-pub struct CreateComponentParams {
-    pub product: String,
-    pub name: String,
-    pub description: String,
-    pub default_assignee: String,
-}
-
-#[derive(Debug, Default, Serialize)]
-#[non_exhaustive]
-pub struct UpdateComponentParams {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub name: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub description: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub default_assignee: Option<String>,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-#[non_exhaustive]
-pub struct Classification {
-    pub id: u64,
-    #[serde(default)]
-    pub name: String,
-    #[serde(default)]
-    pub description: String,
-    #[serde(default)]
-    pub sort_key: u64,
-    #[serde(default)]
-    pub products: Vec<ClassificationProduct>,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-#[non_exhaustive]
-pub struct ClassificationProduct {
-    pub id: u64,
-    #[serde(default)]
-    pub name: String,
-    #[serde(default)]
-    pub description: String,
 }
 
 #[derive(Debug, Clone, Copy, Default, PartialEq, Eq)]
